Changed datatype in OSCAR_DATA_TYPES definition from float64 to float32 for some variables to reduce RAM usage as a first step

+++ 02_Python_Scripts/io_smash.py
@@ -45,10 +45,10 @@ OSCAR_DATA_TYPES = {
     "pdg_mother2": "int32",
     "baryon_number": "int8",
     "strangeness": "int8",
-    "weight": "float64",
-    "partial": "float64",
+    "weight": "float32",
+    "partial": "float32",
     "p_parent_pdg_id": "int32",
-    "block_weight_adj": "float64",
+    "block_weight_adj": "float32",
 }
 
 # -----------------------------
